% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{ACF}
\alias{ACF}
\title{Autocorrelation Function}
\usage{
ACF(X, max_lag = length(X), confidence = 1.96/sqrt(max_lag))
}
\arguments{
\item{X}{time series as a numeric vector.}

\item{max_lag}{integer of the maximum lag to be checked. Set to the length of the time series by default.}

\item{confidence}{that below this threshold covariances are concidered negligible.}
}
\value{
p index of first lag that has dropped below confidence threshold; rho numerical vector of autocorrelation for each lag.
}
\description{
The Autocorrelation Function quantifies the degree of similarity between a time series and its lagged versions over successive time intervals, aiding in the detection of repeating patterns and trends.
}
\examples{
X <- as.numeric(tsdl::tsdl[[1]])
p <- zeitreihen::ACF(X)$p

P <- zeitreihen::simulate(X, 14, zeitreihen::AR, p)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
