% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAX.R
\docType{data}
\name{DAX}
\alias{DAX}
\alias{Date}
\alias{DAX_df}
\alias{DAX_ts}
\title{Financial Data: German Stock Index (DAX)}
\format{
\code{DAX_df} is a data frame with 661 entries (rows) and 2 variables (columns)
named \code{Date} and \code{Close}. The \code{Date} column lists the trading days, while the
\code{Close} column shows the adjusted closing price of the DAX for that date.

\code{DAX_ts} is a time series that shows the adjusted closing price of the DAX
for each trading day.
}
\source{
\href{https://www.finanzen.net/}{Finanzen.net}
}
\usage{
DAX_df
DAX_ts
}
\description{
This dataset contains the stock prices of the German stock index (DAX) for
the period from the beginning of January 2022 to the end of July 2024. The
data is provided both in the form of a data frame (\code{DAX_df}) and as a time
series object (\code{DAX_ts}).
}
\examples{
# Display the first few rows of the data frame
head(DAX_df)

# Plot the stock price
base::plot(DAX_df$Date, DAX_df$Close, type = "l", col = "blue",
    main = "DAX Performance", xlab = "Date", ylab = "Adjusted Closing Price")

}
\keyword{dataset}
