% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLA.R
\name{DLA}
\alias{DLA}
\title{Recursive calculation of coefficients using the Durbin-Levinson Algorithm}
\usage{
DLA(X)
}
\arguments{
\item{X}{A numeric vector representing the time series data. The data should be a stationary time series.}

\item{m}{An integer representing the order of the autoregressive model. The value of \code{m} must satisfy \eqn{0 < m < n},
where \code{n} is the length of the \code{data}.}
}
\value{
A list containing:
\tabular{lcccl}{
\code{phi} \tab \tab \tab \tab A numeric vector of length \code{m}, representing the computed AR coefficients \eqn{\phi{n1}, \dots, \phi{nn}}. \cr
\code{v} \tab \tab \tab \tab A numeric vector of length \code{m}, representing the innovation variances.Row 2, Col 2 \cr
}
}
\description{
\code{DLA} is a function to compute the Durbin-Levinson recursion for time series data.
}
\details{
The Durbin-Levinson algorithm is used to recursively compute the coefficients \eqn{\phi_{n1}, \dots, \phi_{nn}}
for a given time series \eqn{\{x_1, \dots, x_n\}}. The recursion can be expressed as:
\deqn{
\phi_{nn} = \left[ \gamma(n) - \sum_{j=1}^{n-1} \phi_{n-1,j} \gamma(n-j) \right] v_{n-1}^{-1},
}
\deqn{
\left[ \begin{array}{c}
\phi_{n1} \\
\vdots \\
\phi_{n,n-1}
\end{array} \right]
=
\left[ \begin{array}{c}
\phi_{n-1,1} \\
\vdots \\
\phi_{n-1,n-1}
\end{array} \right]
-
\phi_{nn}
\left[ \begin{array}{c}
\phi_{n-1,n-1} \\
\vdots \\
\phi_{n-1,1}
\end{array} \right],
}
\deqn{
v_n = v_{n-1} \left[1 - \phi_{nn}^2\right],
}
where \eqn{\phi_{11} = \frac{\gamma(1)}{\gamma(0)}} and \eqn{v_0 = \gamma(0)}.

The algorithm iteratively calculates these coefficients, which are essential in the analysis of autoregressive processes.
}
\note{
The function includes checks for numerical stability. If any of the calculated variances \eqn{v_n} or the initial
autocovariance value \eqn{\gamma(0)} is too close to zero, the function will stop and return an error to prevent
numerical instability.
}
\examples{
# Example usage
data <- c(2, 4, 6, 8, 10)
model <- DLA(data)
result <- model(2)
print(result)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
