% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IA.R
\name{IA}
\alias{IA}
\title{Compute the Innovations Algorithm for Time Series}
\usage{
IA(X, q = (length(X) - 1), matrix = FALSE)
}
\arguments{
\item{X}{time series as a numeric vector.}

\item{q}{integer of the maximum lag to be checked. Set to the length of the time series - 1 by default.}

\item{matrix}{a boolean wether or not to output a matrix}
}
\value{
A list with two components:
\item{innovations}{A numeric vector containing the computed innovations.}
\item{innovation_variances}{A numeric vector containing the variances of the innovations.}
}
\description{
`IA` can be used to compute the innovations and the innovation variances for a time series.
}
\details{
The innovations algorithm is a recursive method used in time series analysis to compute the best linear predictors of a time series and their associated prediction errors, known as innovations. For a given time series \eqn{\{x_1, \dots, x_n\}}, the algorithm provides a way to decompose the series into uncorrelated components.

The algorithm works by computing the innovations and their variances step by step, using the given autocovariance function of the time series. It is particularly useful in the context of ARMA models and other linear time series models.
}
\examples{
# Generate a sample time series and its autocovariance
x <- rnorm(100)
acf <- acf(x, plot = FALSE)$acf

# Calculate the innovations and their variances
results <- innovations_algorithm(x, acf)
innovations <- results$innovations
variances <- results$innovation_variances

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
