% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make_ma}
\alias{make_ma}
\title{MA(q) generator (`make_ma`)}
\usage{
make_ma(t, theta, sigma = 1)
}
\arguments{
\item{t}{An integer that gives the desired length of the time series.}

\item{theta}{A numeric or complex vector of length \eqn{q} that contains the coefficients of the MA(q) process.}

\item{sigma}{A numeric value representing the variance of the white noise. By default it is set to 1.}
}
\value{
A numeric vector that contains the generated Ma(q) time series.
}
\description{
`make_ma` generates a synthetic time series that follows a MA(q) model.
}
\details{
The MA(q) model can be defined by the equation
\deqn{X_t = Z_t+\theta_1 Z_{t-1} + \theta_2 Z_{t-2} + \dots + \theta_q Z_{t-q}}
where \eqn{t} is the time and \eqn{Z_t} is white 
noise with mean zero and variance \eqn{\sigma^2}.
}
\examples{
Generate a sample MA(1) time series of length 100 with coefficient \eqn{\phi=-0.25}.
ma_one <- make_ma(100,-0.25)
ts.plot(ma_one)
acf(ma_one)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
