% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACF.R
\name{PACF}
\alias{PACF}
\title{Partial Autocorrelation Function (PACF)}
\usage{
PACF(X, max_lag = length(X), confidence = 1.96/sqrt(max_lag))
}
\arguments{
\item{X}{A numeric vector representing the time series data.}

\item{max_lag}{An integer specifying the maximum lag to compute the partial autocorrelation function. By default, it is set to the length of the time series.}

\item{confidence}{A numeric value representing the threshold below which partial autocorrelations are considered negligible. The default value is calculated as \eqn{1.96 / \sqrt{max\_lag}}.}
}
\value{
A list containing the following components:
\item{q}{An integer representing the first lag where the partial autocorrelation drops below the confidence threshold.}
\item{alpha}{A numeric vector containing the partial autocorrelation coefficients for each lag.}
}
\description{
`PACF` computes the partial autocorrelation function of a time series, which quantifies the correlation between the time series and its lagged versions, accounting for the influence of intervening lags. The function also identifies the first lag where the partial autocorrelation drops below a specified confidence threshold.
}
\details{
The partial autocorrelation function (PACF) is used in time series analysis to measure the correlation between the time series and its lagged versions, after removing the effect of shorter lags. This makes PACF particularly useful in identifying the direct effect of a lag on the time series, which is helpful in determining the order of an autoregressive (AR) model.

For a given time series \eqn{\{X_1, \dots, X_n\}}, the PACF is computed for lags up to `max_lag` by solving the Yule-Walker equations. The function also identifies the first lag where the partial autocorrelation coefficient falls below a specified confidence threshold, returning this lag as `q`.
}
\examples{
# Generate a sample time series
X <- rnorm(100)

# Calculate the partial autocorrelation function and the first significant lag
result <- PACF(X)
q <- result$q
alpha <- result$alpha

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
