% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Estimate the Spectral Density of a Time Series using a Periodogram}
\usage{
periodogram(x, lambda)
}
\arguments{
\item{x}{A numeric or complex atomic vector representing the time series data. The series should be stationary,
meaning that its mean and variance do not change over time.}

\item{lambda}{A numerical value from the interval \eqn{(- \pi, \pi]}, which represents the frequency at which the periodogram is evaluated.}
}
\value{
A numerical value that represents the periodogram.
}
\description{
`periodogram` computes the periodogram of a time series, which can be used to estimate the spectral density.
}
\details{
The periodogram is a sample-based function from which we obtain estimators of the spectral density.
For a given time series \eqn{\{x_1, \dots, x_n\}}, the periodogram can be computed using the formula:

\deqn{I_n(\lambda) = \displaystyle{\frac{1}{n}} \left| \displaystyle{\sum_{t=1}^n x_t e^{-it\lambda}} \right|^2.}
}
\examples{
# Generate a sample time series
x <- rnorm(100)

# Calculate the periodogram
periodogram(x, pi / 2)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
