% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Custom Time Series Plotting Function}
\usage{
plot(X, samples)
}
\arguments{
\item{X}{A numeric vector representing the time series data, including the predictions.}

\item{samples}{an integer specifying the number of samples to evaluate the periodogram with.}
}
\value{
A line plot displaying the spectral density estimate of the time series.
}
\description{
\code{plot} generates a custom line plot of a time series vector, highlighting the last \code{h} values as predictions. The plot displays the time series in blue and the predictions in red, with an appropriate legend.
}
\details{
This function is designed to visualize a time series, where the last \code{h} values are considered predictions and are highlighted in a different color. The main part of the time series is plotted in blue, while the predicted values are plotted in red. The function automatically generates a legend to distinguish between the observed and predicted values.
}
\examples{
# Generate a random time series with 100 elements
X <- rnorm(100)

# Plot the time series with the last 10 elements highlighted as predictions
plot(X, 100)

}
