% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ACVF.R
\name{sample_ACVF}
\alias{sample_ACVF}
\title{The Sample Autocovariance Function}
\usage{
sample_ACVF(X, h = 0:(length(X) - 1))
}
\arguments{
\item{X}{A numeric or complex atomic vector representing the time series data.}

\item{h}{An integer atomic vector representing the lag values, where each value \eqn{h_i} must be unique and fulfill the condition: \eqn{-\text{length}(X) < h_i < \text{length}(X)}. By default, it is set to the intervall \eqn{[0, \text{length}(X) - 1]}.}
}
\value{
A numeric atomic vector containing the calculated values of the sample autocovariance function corresponding to each entry in \code{h}.
}
\description{
\code{sample_ACVF} computes the sample autocovariance of a given time series at a specified lag.
}
\details{
Let \eqn{\{x_1, \dots , x_n \}} be observations of a time series. The \strong{sample mean} of \eqn{x_1, \dots , x_n} is

\deqn{\bar{x} = \displaystyle{\frac{1}{n} \sum^n_{t=1} x_t}.}

The \strong{sample autocovariance function} measures the covariance between values of a time series across different lags and is given by

\deqn{\hat{\gamma}(h) := n^{-1} \displaystyle{\sum_{t=1}^{n - |h|} (x_{t + |h|} - \bar{x}) \cdot (x_t - \bar{x})}, \quad -n < h < n.}
}
\examples{
# Example 1: Basic Usage
X <- c(1, 2, 3, 4, 5)
h <- c(0, 1, 2)
sample_ACVF(X, h)

# Example 2: Sample ACVF with a Single Lag
X <- c(2, 4, 6, 8, 10)
h <- 0
sample_ACVF(X, h)

# Example 3: Sample ACVF with Random Data
X <- rnorm(100)
h <- c(-2, 0, 2)
sample_ACVF(X, h)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
