% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Time Series Simulation Function}
\usage{
simulate(X, h = 1, model, lag, coeffs, plot = False)
}
\arguments{
\item{X}{A numeric vector representing the time series data.}

\item{h}{an integer specifying the number of predictions to be computed. The default value is \code{1}.}

\item{model}{A function from the \code{zeitreihen} package to be used as the predictive model. The default is \code{zeitreihen::auto_model}.}

\item{lag}{An integer specifying the lag order for the model. This parameter can be used to override the default lag of the chosen model.}

\item{coeffs}{A numeric vector specifying the coefficients for the model. This parameter can be used to override the default coefficients of the chosen model.}

\item{plot}{A logical value indicating whether to plot the time series along with its predictions. The default value is \code{FALSE}.}
}
\value{
A numeric vector containing the \code{h} predicted values appended to the original time series.
}
\description{
\code{simulate} generates future values (predictions) for a given time series by applying a specified model. The function can also plot the original series with the predictions appended if desired.
}
\details{
This function simulates future observations of a time series by using a specified model from the \code{zeitreihen} package. The user can specify the number of predictions (\code{h}), the model to be used, the lag order, and the model coefficients. If \code{plot} is set to \code{TRUE}, the function will generate a plot showing the time series with the predictions highlighted.
}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
