% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempDE.R
\docType{data}
\name{tempDE}
\alias{tempDE}
\alias{Year}
\alias{tempDE_df}
\alias{tempDE_ts}
\title{Weather Data: Average Annual Temperature in Germany}
\format{
\code{tempDE_df} is a data frame with 143 entries (rows) and 2 variables (columns)
named \code{Year} and \code{Temperature}. The \code{Year} column contains the respective
years, while the \code{Temperature} column gives the average annual temperature in
the whole of Germany.

\code{tempDE_ts} is a time series that shows the average annual temperature in
Germany for each year.
}
\source{
\href{https://www.dwd.de/}{Deutscher Wetterdienst}
}
\usage{
tempDE_df
tempDE_ts
}
\description{
This dataset contains the average annual temperatures in Germany from 1881 to
2023. The data is provided both as a data frame (\code{tempDE_df}) and as a time
series object (\code{tempDE_ts}).
}
\examples{
# Display the first few rows of the data frame
head(tempDE_df)

# Plot the temperature
base::plot(tempDE_df$Year, tempDE_df$Temperature, type = "p", col = "blue",
    main = "Average Annual Temperature in Germany (1881-2023)",
    xlab = "Year", ylab = "Temperature (°C)")
lines(tempDE_df$Year, tempDE_df$Temperature, col = "red")

}
\keyword{dataset}
