\name{zeitreihen-package}
\alias{zeitreihen-package}
\alias{zeitreihen}
\docType{package}
\title{
\packageTitle{zeitreihen}
}
\description{
\packageDescription{zeitreihen}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{zeitreihen}

Maintainer: \packageMaintainer{zeitreihen}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = zeitreihen)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{zeitreihen}}
%\section{Documentation Index}{\packageIndices{zeitreihen}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
