% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{ACF}
\alias{ACF}
\title{Autocorrelation Function (ACF)}
\usage{
ACF(X, max_lag = length(X), confidence = 1.96/sqrt(max_lag))
}
\arguments{
\item{X}{A numeric vector representing the time series data.}

\item{max_lag}{An integer specifying the maximum lag to compute the autocorrelation function. By default, it is set to the length of the time series.}

\item{confidence}{A numeric value representing the threshold below which autocorrelations are considered negligible. The default value is calculated as \eqn{1.96 / \sqrt{max\_lag}}.}
}
\value{
A list containing the following components:
\item{p}{An integer representing the first lag where the autocorrelation drops below the confidence threshold.}
\item{rho}{A numeric vector containing the autocorrelation values for each lag.}
\item{kappa}{A numeric value representing the autocovariance at the last computed lag.}
\item{sigma2}{A numeric value representing the variance of the time series.}
}
\description{
`ACF` computes the autocorrelation function of a time series, which measures the correlation between the time series and its lagged versions over successive time intervals. The function also identifies the first lag where the autocorrelation drops below a specified confidence threshold.
}
\details{
The autocorrelation function (ACF) is a key tool in time series analysis that helps to identify repeating patterns, trends, or seasonal effects by quantifying the degree of similarity between a time series and its lagged versions. Given a time series \eqn{\{X_1, \dots, X_n\}}, the function calculates the autocorrelation for lags ranging from 0 up to `max_lag`. The function also checks for the first lag where the autocorrelation falls below a specified confidence threshold, returning this lag as `p`.
}
\examples{
# Generate a sample time series
X <- rnorm(100)

# Calculate the autocorrelation function and the first significant lag
result <- ACF(X)
p <- result$p
rho <- result$rho

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
