% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAX.R
\docType{data}
\name{DAX}
\alias{DAX}
\alias{date}
\alias{DAX_df}
\alias{DAX_ts}
\title{Financial Data: German Stock Index (DAX)}
\format{
\code{DAX_df} is a data frame with 661 entries (rows) and 2 variables (columns)
named \code{data} and \code{value}.
}
\usage{
DAX_df
DAX_ts
}
\description{
This file contains the stock prices of the German stock index (DAX) for the
period from January 2022 to July 2024. The data is provided both in the form
of a data frame and as a time series.
}
\keyword{datasets}
