% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLA.R
\name{DLA}
\alias{DLA}
\title{Recursive Calculation of Coefficients using the Durbin-Levinson Algorithm}
\usage{
DLA(X)
}
\arguments{
\item{X}{A numeric or complex atomic vector representing the time series data.
The series should be stationary, meaning that its mean and variance do not change over time.}
}
\value{
A list with two components:
\item{phi}{A numeric atomic vector representing the calculated AR coefficients \eqn{\phi_{n1}, \dots, \phi_{nn}}.}
\item{nu}{A numeric atomic vector representing mean squared error of the one step predictors \eqn{\nu_{i} = E(X_{i+1} - (\theta_{i1}X_{n} + ... + \theta_{ii}X_i))^{2}}, i > 0 and \eqn{\nu_0 = \gamma(0)}.}
}
\description{
\code{DLA} can be used to perform the Durbin-Levinson recursion on time series data.
}
\details{
The \strong{Durbin-Levinson Algorithm} is used to recursively calculate the coefficients \eqn{\phi_{n1}, \dots, \phi_{nn}}
for a given time series \eqn{\{x_1, \dots, x_n\}}. The recursion can be expressed as follows:

\deqn{\phi_{nn} = \left[ \gamma(n) - \displaystyle{\sum_{j=1}^{n-1} \phi_{n-1, j} \gamma(n-j)} \right] \nu_{n-1}^{-1},}

\deqn{
\left[ \begin{array}{c} 
       \phi_{n1} \\
       \vdots \\
       \phi_{n, n-1}
       \end{array}
\right]
=
\left[ \begin{array}{c}
       \phi_{n-1,1} \\
       \vdots \\
       \phi_{n-1,n-1}
       \end{array}
\right]
-
\phi_{nn}
\left[ \begin{array}{c}
       \phi_{n-1,n-1} \\
       \vdots \\
       \phi_{n-1,1}
       \end{array}
\right]}

and

\deqn{\nu_n = \nu_{n-1} \left[1 - \phi_{nn}^2\right],}

where \eqn{\phi_{11} = \gamma(1)/\gamma(0)} and \eqn{\nu_0 = \gamma(0)}.

The algorithm iteratively calculates these coefficients, which are essential for the analysis of autoregressive processes
and can be used as inputs of the \code{\link{make_AR}} function.
}
\note{
The function includes checks for numerical stability. If any of the calculated variances \eqn{\nu_n} or the initial
autocovariance value \eqn{\gamma(0)} is too close to zero, the function terminates and returns an error to prevent numerical instability.
}
\examples{
# Example Usage
data <- c(2, 4, 6, 8, 10)
model <- DLA(data)
result <- model(2)
print(result)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
