% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_AR.R
\name{make_AR}
\alias{make_AR}
\title{Generator for AR(p) processes}
\usage{
make_AR(t, phi, sigma = 1, start = numeric(length(phi)))
}
\arguments{
\item{t}{An integer specifying the desired length of the time series.}

\item{phi}{A numeric or complex atomic vector of length \code{p} containing the coefficients of the AR(p) process.}

\item{sigma}{A numeric value representing the standard deviation of the white noise, with a default value of 1.}

\item{start}{A numeric atomic vector containing the first \code{p} observations, which are 0 by default.}
}
\value{
A numeric atomic vector containing the generated AR(p) time series.
}
\description{
\code{make_AR} generates a synthetic time series that follows an AR(p) model.
}
\details{
The AR(p) model can be defined by the equation

\deqn{X_t = \phi_1 X_{t-1} + \phi_2 X_{t-2} + \dots + \phi_p X_{t-p} + Z_t,}

where \eqn{X_t} represents the time series at time \eqn{t} and \eqn{Z_t} is white noise with a mean of zero and a variance of \eqn{\sigma^2}.
}
\examples{
# Example 1: Simple AR(1) model
ar_one <- make_AR(t = 10, phi = -0.7, sigma = 9, start = 7)
graphics::plot(ar_one, type = "l")

# Example 2: Simple AR(2) model
ar_two <- make_AR(t = 10, phi = c(1 + 1i, 0))
graphics::plot(ar_two, type = "l")

# Example 3: Simple AR(3) model
ar_three <- make_AR(t = 10, phi = c(6, 9, 1))
graphics::plot(ar_three, type = "l")

# Example 4: Simple AR(4) model
ar_four <- make_AR(t = 10, phi = c(0, 4, 1, 2), start = c(9, 5))
graphics::plot(ar_four, type = "l")

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
