% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MA.R
\name{make_MA}
\alias{make_MA}
\title{Generator for MA(q) processes}
\usage{
make_MA(t, theta, sigma = 1)
}
\arguments{
\item{t}{An integer specifying the desired length of the time series.}

\item{theta}{A numeric or complex atomic vector of length \code{q} containing the coefficients of the MA(q) process.}

\item{sigma}{A numeric value representing the standard deviation of the white noise, with a default value of 1.}
}
\value{
A numeric atomic vector containing the generated MA(q) time series.
}
\description{
\code{make_MA} generates a synthetic time series that follows an MA(q) model.
}
\details{
The MA(q) model can be defined by the equation

\deqn{X_t = Z_t + \theta_1 Z_{t-1} + \theta_2 Z_{t-2} + \dots + \theta_q Z_{t-q},}

where \eqn{X_t} represents the time series at time \eqn{t} and \eqn{Z_t} is white noise with a mean of zero and variance of \eqn{\sigma^2}.
}
\examples{
# Example 1: Simple MA(1) model
ma_one <- make_MA(t = 10, theta = -0.25, sigma = 2)
graphics::plot(ma_one, type = "l")

# Example 2: Simple MA(2) model
ma_two <- make_MA(t = 10, theta = c(9, 2))
graphics::plot(ma_two, type = "l")

# Example 3: Simple MA(3) model
ma_three <- make_MA(t = 10, theta = c(1, 3, 4))
graphics::plot(ma_three, type = "l")

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
