% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACF.R
\name{PACF}
\alias{PACF}
\title{Partial Autocorrelation Function}
\usage{
PACF(X, max_lag = length(X), confidence = 1.96/sqrt(max_lag))
}
\arguments{
\item{X}{time series as a numeric vector.}

\item{max_lag}{integer of the maximum lag to be checked. Set to the length of the time series by default.}

\item{confidence}{that below this threshold covariances are concidered negligible.}
}
\value{
q index of first lag that has dropped below confidence threshold; alpha numerical vector of autocorrelation for each lag.
}
\description{
The Partial Autocorrelation Function measures the correlation between a time series and its lagged values, to identify the direct effect of a lag on the series.
}
\examples{
X <- as.numeric(tsdl::tsdl[[1]])
q <- zeitreihen::PACF(X)$q

P <- zeitreihen::simulate(X, 14, zeitreihen::MA, q)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
