% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Custom Spectral Density Plotting Function}
\usage{
plot(X, n)
}
\arguments{
\item{X}{A numeric vector representing the time series data, including the predictions.}

\item{n}{an integer specifying the number of samples to evaluate the periodogram with.}
}
\value{
A line plot displaying the spectral density estimate of the time series.
}
\description{
\code{plot} computes periodogram values, plots the spectral density, and includes a legend. It samples over \eqn{(-\pi, \pi]} and manages errors.
}
\details{
The function computes periodogram values using the \code{periodogram} function from the \code{stats} package. It then generates a plot of the spectral density estimate with the \code{plot} function from the \code{graphics} package, including a legend to distinguish between observed and predicted values. The samples are evaluated at evenly spaced points over the interval \eqn{(-\pi, \pi]}. Additionally, the function incorporates error handling to manage issues such as invalid inputs or missing packages.
}
\examples{
# Generate a random time series with 100 elements
X <- rnorm(500)

# Plot the spectral density
zeitreihen::plot(X, 100)

}
