% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Custom Time Series Plotting Function}
\usage{
plot(X_hat, h)
}
\arguments{
\item{h}{integer specifying the number of predictions in X.}

\item{X}{time series as a numeric vector.}
}
\value{
plot of X with `h` predictions marked in red.
}
\description{
This function plots a vector `X` as a time series, with the last `h` indicated to be the predictions.
}
\examples{
X <- rnorm(100) # Generate a random vector of 100 elements
zeitreihen::plot(X, 10) # Plot with the last 10 elements in a different color

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
