% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ACVF.R
\name{sample_ACVF}
\alias{sample_ACVF}
\title{sample autocovariance function}
\usage{
sample_ACVF(x, h = 0:(length(x) - 1))
}
\arguments{
\item{x}{A numeric vector representing Time Series data. NA entries are not allowed.}

\item{h}{A integer vector contains the lag. NA entries are not allowed and h should be unique.

For all entries of h: -length(x) < entries < length(x).}
}
\value{
A numeric vector, which contains computation of the sample autocovariance function based on the entries of h.
}
\description{
Computes the sample autocovariance of the lag h based on data x.
}
\details{
Consider the observed data \eqn{\{x_1,x_2,...,x_n\}}, which represents a Time Series.
The sample autocovariance function (\eqn{\hat \gamma}) measures the covariance between values of a Time Series over different lag.

\eqn{\hat \gamma(h) = \sum_{t=1}^{n - |h|} (x_{t + |h|)}- \bar x) \cdot (x_t - \bar x)}, -n < h < n and
\eqn{\bar x = n^{-1} \cdot \sum_{t=1}^{n} x_{t}}
}
\examples{
x <- c(1, 2, 3, 4)
h <- c(0, 1, 2)
sample_ACVF(x, h)
}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
