% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulation Function for Timeseries}
\usage{
simulate(X, h = 1, model = zeitreihen::auto_model, lag, coeffs, plot = False)
}
\arguments{
\item{X}{time series as a numeric vector.}

\item{h}{integer specifying the number of predictions to be computed. Set to `1` by default.}

\item{model}{function of the zeitreihen package to use as a model. Set to `zeitreihen::auto_model` by default.}

\item{lag}{optionally overwrite default model lag to compute for time series.}

\item{coeffs}{optionaly overwrite default model coeffs to compute for time series.}

\item{plot}{optionally output a plot of the time series and its predictions. Set to `False` by default.}
}
\value{
`P` predictions as a numerical vector
}
\description{
This function starts the simulation process, provided a time series `X` it returns `h`-amount of predictions `P`.
}
\examples{
X <- as.numeric(tsdl::tsdl[[1]])
# minimal function call with 31 predictions
zeitreihen::simulate(X, 31)
# output a plot
zeitreihen::simulate(X, 31, plot = True)
# specify AR model and force p = 1
zeitreihen::simulate(X, 14, zeitreihen::AR, 1)

}
\references{
Brockwell, P.J., Davis, R.A. (2016) \emph{Introduction to Time Series and Forecasting}. Springer.
}
