\name{zeitreihen-package}
\alias{zeitreihen-package}
\alias{zeitreihen}
\docType{package}
\title{
\packageTitle{zeitreihen}
}
\description{
\packageDescription{zeitreihen}
}
\details{
  zeitreihen (german for time series) is a package for time series analasis and comprises methods for analysing data in order to extract meaningful statistis and other characterisitcs of the data such as time series forecasting.
}
\author{
\packageAuthor{zeitreihen}

Maintainer: \packageMaintainer{zeitreihen}
}
\references{
  Introduction to time series and forecasting, Springer (2016), https://link.springer.com/book/10.1007/978-3-319-29854-2
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = zeitreihen)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{zeitreihen}}
%\section{Documentation Index}{\packageIndices{zeitreihen}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
  %%  ~~ Optional simple examples of the most important functions ~~
}
